+++ tests/Stubs/IndividualScopeControllerStub.php
@@ -2,10 +2,13 @@
 
 namespace Dingo\Api\Tests\Stubs;
 
-use Dingo\Api\Routing\Controller;
+use Illuminate\Routing\Controller;
+use Dingo\Api\Routing\ControllerTrait;
 
 class IndividualScopeControllerStub extends Controller
 {
+    use ControllerTrait;
+
     public function __construct()
     {
         $this->scope(['foo', 'bar'], 'index');